# Avoid extra call of safe_path in file_extension()