This speeds up `audiofile.read()` and all the functions from `audiofile.core.info` by not calling `audeer.safe_path()` when getting the file extension of a file, as we called `audeer.safe_path(file)` before.