@hagenw 2024-01-15
Not with the [benchmarks in the docs](https://audeering.github.io/audiofile/benchmark.html) as there we look only at 160 files. But it can be measured when using more files, e.g.

Execution time when reading 100,000 files with a duration of 0.1 s and a sampling rate of 8000 Hz.

| main | this branch |
| --- | --- |
| 15.2 s | 11.5 s |

<details><summary>Benchmark code</summary>

```python
import os
import time

import numpy as np

import audeer
import audiofile


duration = 0.1
sampling_rate = 8000
repetitions = 100000

# Create files for benchmark
folder = audeer.mkdir('mix')
files = [
    audeer.path(folder, f'long-{n}.wav')
    for n in range(repetitions)
]
for file in files:
    if not os.path.exists(file):
        signal = np.random.normal(0, 0.1, (1, int(sampling_rate * duration)))
        signal /= (np.max(np.abs(signal)) + 0.0001)
        audiofile.write(file, signal, sampling_rate)

# benchmark
start = time.time()
for file in files:
    audiofile.read(file)
end = time.time()
print(end - start)
```

</details>

I also saw this when profiling `auglib`, which spent 1 s on `audeer.safe_path()` and half of it inside `file_extension()`.