@frankenjoe 2024-01-15
Wow, I was expecting a call to `audeer.path()` to not have a measurable impact. At least not if the path is absolute and does not contain relative sub-paths. Since we are using this function all over the place, maybe we could add such a check to `audeer.path()` then?