@hagenw 2024-01-15
Yes, it has a big impact if you loop over lot of files and all your other processing is relatively fast. E.g. when removing `audeer.safe_path()` completely when executing the benchmark above, the execution time is reduced to 8.0 s.
The expensive part inside `audeer.path()` is `os.path.realpath()` which translates symbolic links to the actual path.

One of the advantages of using `os.path.realpath()` is the following:

```python
import audeer
import os

audeer.touch('test.wav')
os.symlink('test.wav', 'test.mp3')

audeer.file_extension('test.mp3')
```
returns
```
'wav'
```