@frankenjoe 2024-01-15
Ok, I see. Basically, it means there is at least one disk operation we cannot avoid since we cannot derive from the string if a path is symbolic or not.