@hagenw 2024-01-15
I added https://github.com/audeering/audeer/issues/131 for further discussion on the `audeer.path()` issue as this is not relevant to the proposed changes of this pull request.